rdmd: add unittested implementation of -o argument parsing
The two new functions do a similar job to the existing `dashOh` callback
but split the work into two easily testable parts:

  * a method of the `RDMDArgs` struct, which can be used directly
    as a callback with `getopt`, and which ensures the results are
    written into appropriate struct fields

  * a standalone implementation of the underlying parsing, which
    accepts arbitrary string and bool parameters into which to
    write the parsed values

Unittests are provided for both, but the latter offers an easier route
to more general testing.

Unlike `dashOh`, this new implementation does not write `-od` and `-of`
output into the same string: instead, it is expected that a different
string will be used to store the results of each.  No more than one of
each is supported, but the question of whether both `-of` and `-od` can
be supplied together is left for the caller to decide (i.e. parsing of
`-o...` arguments is considered separately from their validation).